Idan CR: pydantic dataclasses + seperate Schema file
